Fix issue Font not found from HTMLLabel [Windows]
The initial problem is that the `load` function found in font\__init__.py
can accept a list of names. This is what happens with the HTMLLabel
font tag. The function tries to find each font in the list with the code
    # Find first matching name
    if type(name) in (tuple, list):
        for n in name:
            if _font_class.have_font(n):
                name = n
                break
        else:
            name = None

The problem is that GDIPlusFont which derives from GDIFont only look in the
system fonts collection. But if we have added some private fonts, it's not
looking there.

This fix add a `have_font` class method to GDIPlusFont which first look for the
font in its private font collection. Only if it does not find it, it calls
its parent class `have_font` class method.